requirements.txt
bs4
pandas
gradio
transformers
accelerate
torch>=2.0.0
google-api-python-client
google-auth-httplib2
google-auth-oauthlib    